update folder structure, move files, update package names
