+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/MGTheTrain/terraform-provider-blob-services/src
+module github.com/MGTheTrain/terraform-provider-blob-services
 
 go 1.21.3
